Shovelnose: Multiple Nightmare instances and refactoring
This commit introduces some additions as well as refactoring to make
Shovelnose a bit clearer and to add slight optimizations. In common, the error module was added, which
outputs red text to the console. This was added to help pick out errors
from regular debug output.

In Shovelnose, the option to specify multiple Nightmare instances was
added to speed things up. The search function now uses objects to
cleanly seperate instances and keep variables common to the object.
This reduces the memory footprint of the recursive search
implementation.

Warning - Using multiple SN instances is buggy. Still commiting for
remote backup.